added verification and authorization checks for the current user

+++ routers/score.py
@@ -5,10 +5,9 @@ from fastapi import FastAPI, APIRouter, HTTPException, status, Depends
 from typing import Annotated
 from sqlalchemy import select
 from sqlalchemy.orm import Session
-
+from auth import CurrentUser
 
 router = APIRouter()
-
 
 @router.get('/{score_id}', response_model=ScoreResponse)
 def get_scores(score_id : int, db: Annotated[Session, Depends(get_db)]):
@@ -22,33 +21,32 @@ def get_scores(score_id : int, db: Annotated[Session, Depends(get_db)]):
 
 
 @router.post('', response_model=ScoreCreate, status_code=status.HTTP_201_CREATED)
-def create_score(score : ScoreCreate, db: Annotated[Session, Depends(get_db)]):
-    result = db.execute(select(models.User).where(models.User.id == score.user_id))
-    existing_user = result.scalars().first()
+def create_score(score : ScoreCreate, current_user: CurrentUser, db: Annotated[Session, Depends(get_db)]):
 
-    if not existing_user:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="No User Found")
-    else:
-        new_score = models.Score(
-            score=score.score,
-            user_id=score.user_id,
-            scale=score.scale,
-            scale_key=score.scale_key
-        )
+    new_score = models.Score(
+        score=score.score,
+        user_id=current_user.id,
+        scale=score.scale,
+        scale_key=score.scale_key
+    )
 
-        db.add(new_score)
-        db.commit()
-        db.refresh(new_score)
-        return new_score
+    db.add(new_score)
+    db.commit()
+    db.refresh(new_score)
+    return new_score
     
+
 @router.delete('/{score_id}',
                status_code=status.HTTP_204_NO_CONTENT)
-def delete_score(score_id: int, db: Annotated[Session, Depends(get_db)]):
+def delete_score(score_id: int, current_user : CurrentUser, db: Annotated[Session, Depends(get_db)]):
     result = db.execute(select(models.Score).where(models.Score.id == score_id))
     score = result.scalars().first()
 
     if not score:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
-    else:
-        db.delete(score)
-        db.commit()
+    
+    if score.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to delete this score")
+    
+    db.delete(score)
+    db.commit()
